leadflow_ai/requirements.txt
fastapi
uvicorn
langchain
langgraph
openai
pinecone-client
supabase-py
pydantic
python-dotenv
httpx
requests
aiohttp
tiktoken 